refactor(auth): move the authentication layer into src/core/auth
Relocate the realm-agnostic parts of the auth stack (csrf, cookies,
token transport, redis keys/scripts, jwt payload schema, session token
issuance/rotation, and the shared error classes) from src/user/auth
into src/core/auth so a second class of principals can reuse them
without copying. Pure move: no signature, Redis key, or cookie name
changes.

- src/core/auth/tokens.py takes over create_access_token,
  create_refresh_token and rotate_refresh_token (issue_token replaces
  the former private _issue_token).
- src/core/auth/errors.py takes PermissionDeniedError, CsrfFailedError,
  TokenExpiredError and InvalidCredentialsError; src/user/auth/errors.py
  keeps only UserBlockedError and UserNotVerifiedError.
- src/user/auth/security.py keeps the one-time verification/reset-password
  token flow, absorbing store/validate/invalidate_active_one_time_token
  from token_helpers.py since they depend on normalized email, not on
  session state.
