+++ src/user/usecases/update_password.py
@@ -6,6 +6,8 @@ from fastapi import Depends
 from redis.asyncio import Redis
 
 from loggers import get_logger
+from src.core.auth.errors import InvalidCredentialsError
+from src.core.auth.token_helpers import invalidate_all_user_sessions
 from src.core.cache.dependencies import get_cache
 from src.core.cache.interface import Cache
 from src.core.database.session import get_unit_of_work
@@ -17,9 +19,7 @@ from src.core.errors.exceptions import (
 from src.core.redis.dependencies import get_redis_client
 from src.core.schemas import SuccessResponse
 from src.core.utils.security import hash_password, mask_email, verify_password
-from src.user.auth.errors import InvalidCredentialsError
 from src.user.auth.schemas import UserNewPassword
-from src.user.auth.token_helpers import invalidate_all_user_sessions
 from src.user.cache_keys import user_cache_keys
 
 logger = get_logger(__name__)
