+++ tests/unit/src/user/auth/test_csrf.py
@@ -1,4 +1,4 @@
-from src.user.auth.csrf import build_csrf_token, verify_csrf_token
+from src.core.auth.csrf import build_csrf_token, verify_csrf_token
 
 SECRET = "unit-test-csrf-secret-key-value-32"
 REFRESH_TOKEN = "header.payload.signature"
