+++ tests/unit/src/user/auth/test_errors.py
@@ -1,5 +1,6 @@
 import pytest
 
+from src.core.auth import errors as core_errors
 from src.core.errors.codes import ErrorCode
 from src.user.auth import errors
 
@@ -9,10 +10,10 @@ from src.user.auth import errors
     [
         (errors.UserBlockedError, 403, ErrorCode.USER_BLOCKED),
         (errors.UserNotVerifiedError, 403, ErrorCode.USER_NOT_VERIFIED),
-        (errors.PermissionDeniedError, 403, ErrorCode.PERMISSION_DENIED),
-        (errors.CsrfFailedError, 403, ErrorCode.CSRF_FAILED),
-        (errors.TokenExpiredError, 401, ErrorCode.TOKEN_EXPIRED),
-        (errors.InvalidCredentialsError, 400, ErrorCode.INVALID_CREDENTIALS),
+        (core_errors.PermissionDeniedError, 403, ErrorCode.PERMISSION_DENIED),
+        (core_errors.CsrfFailedError, 403, ErrorCode.CSRF_FAILED),
+        (core_errors.TokenExpiredError, 401, ErrorCode.TOKEN_EXPIRED),
+        (core_errors.InvalidCredentialsError, 400, ErrorCode.INVALID_CREDENTIALS),
     ],
 )
 def test_domain_error_declares_status_and_code(error_class, status, code) -> None:
@@ -21,4 +22,4 @@ def test_domain_error_declares_status_and_code(error_class, status, code) -> Non
 
 
 def test_invalid_credentials_logs_at_debug() -> None:
-    assert errors.InvalidCredentialsError.log_level == "debug"
+    assert core_errors.InvalidCredentialsError.log_level == "debug"
